Add 'my_inject' method to 'Enumerable' module
Performs specific operation, given by block, on each element in an
array and returns a single result. The method takes an
optional initial operand, which will be used as the starting value for
the operation if provided. If no initial operand is given,
the first element in the array will be used as the starting value.